First draft of a minigame to test out the engine
This commit is mostly for you to test if the project works on your side,
i'll add documentation on file/object logic later on the wiki, as well
as comments in the code for better understanding.

The rules are simple: you move left and right with the arrow keys, press
A to shoot and kill enemies. When enemies die, the score goes up and
they release 4 projectiles, if those projectiles hit enemies, the score
goes up with a multiplier, the bigger the chain reaction, the bigger the
multiplier. Press Enter to return to the menu.

Have fun